refactor: added getPolyfills for better testing
